Move Text → Original and Transliteration → Text.
With original texts in the Text file there is a discrepancy in
orthographies in the Text files. I’ve discarded the “Transliteration”
type and moved the texts not in RP orthography to a new “Original”
type. The transliterated files are now Text type. This means you can
search all Text type files with the same orthographic expectations.