Finished Johnson’s Strong Man.
Completed transliteration of Dauenhauer’s transcription. Nora was not
familiar with Southern Tlingit and probably “Northernized” many things
as well as not hearing falling tone. The audio recording is available
and should be reviewed to update this transcription. Users will need to
check the dialect since spelling is markedly different especially with
the missing classifier /i/ vowel. High tone on prefixes is not written
here; Leer (1978, 1991, 2001) describes this as predictable. We may not
want to write falling before sonorants like héèn since it is
predictable, but falling elsewhere like ḵáà should be written.