Fixing up translation format.
Starting to finalize the format of glosses. The translation line is
contained within ‹ and › so that it can be distinguished from the gloss
lines which have no marking; the text line starts with a number. At
present there is only one gloss line, but there could be a few
eventually, with different levels of segmentation. I originally had the
translation line with ‘…’ but that conflicts with the use of ‘ and ’
for punctuation, particularly since quoted speech can take up multiple
lines. If ‹ and › don’t work we can change them later.