Couple of typos.
At natí in line 287 is a nominalization so write it as one word.

The wéitʼaat in line 285 could either be wéitʼaa-t or a typo for
wéitʼát. I’m going with the latter because the PP as a location of
existence doesn’t really make sense here. It could also be a typo for
wéitʼaa, so the original recording should be checked.