De-gender clan name
The «… sháa» ‘woman of’ construction is often used for female clan members, as opposed to a generic form (–.ádi ~ —.eidí ‘ones of’) which is not inherently gendered for male (and other gender) clan members. The feminine gender form makes it more complicated to search for clan names, so for the corpus we use a clan name’s generic form for everyone.